refactor(presentation): update controllers to use entities and shared errors
