+++ src/presentation/controllers/choose-question-best-answer/choose-question-best-answer.controller.ts
@@ -2,10 +2,10 @@ import type { WebController } from '@/core/presentation/web-controller'
 import type { UseCase } from '@/core/application/use-case'
 import { JWTService } from '@/infra/auth/jwt/jwt-service'
 import type { HttpRequest, HttpResponse } from '@/infra/http/ports/http-protocol'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { forbidden, notFound, ok } from '@/presentation/helpers/http-helpers'
-import { extractToken } from '@/util/auth/extract-token'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { extractToken } from '@/shared/util/auth/extract-token'
 
 export class ChooseQuestionBestAnswerController implements WebController {
   constructor (private readonly chooseQuestionBestAnswerUseCase: UseCase) {}
