+++ src/presentation/controllers/comment-on-question/comment-on-question.controller.ts
@@ -2,9 +2,9 @@ import type { WebController } from '@/core/presentation/web-controller'
 import type { UseCase } from '@/core/application/use-case'
 import { JWTService } from '@/infra/auth/jwt/jwt-service'
 import type { HttpRequest, HttpResponse } from '@/infra/http/ports/http-protocol'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { created, notFound } from '@/presentation/helpers/http-helpers'
-import { extractToken } from '@/util/auth/extract-token'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { extractToken } from '@/shared/util/auth/extract-token'
 
 export class CommentOnQuestionController implements WebController {
   constructor (private readonly commentOnQuestionUseCase: UseCase) {}
