+++ src/presentation/controllers/delete-comment/delete-comment.controller.test.ts
@@ -1,9 +1,9 @@
 import type { UseCase } from '@/core/application/use-case'
 import { JWTService } from '@/infra/auth/jwt/jwt-service'
 import { UseCaseStub } from '@/infra/doubles/use-case.stub'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { DeleteCommentController } from './delete-comment.controller'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 vi.mock('@/lib/env', () => ({
   env: {
