+++ src/presentation/controllers/fetch-users/fetch-users.controller.test.ts
@@ -1,7 +1,7 @@
 import type { UsersRepository } from '@/application/repositories/users.repository'
 import { InMemoryUsersRepository } from '@/infra/persistence/repositories/in-memory/in-memory-users.repository'
-import { makeUser } from '@/util/factories/domain/make-user'
 import { FetchUsersController } from './fetch-users.controller'
+import { makeUser } from '@/shared/util/factories/domain/make-user'
 
 function makePaginatedResponse<T> (
   page: number,
