+++ src/presentation/controllers/get-user-by-email/get-user-by-email.controller.ts
@@ -1,8 +1,8 @@
 import type { WebController } from '@/core/presentation/web-controller'
 import type { UseCase } from '@/core/application/use-case'
 import type { HttpRequest, HttpResponse } from '@/infra/http/ports/http-protocol'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { notFound, ok } from '@/presentation/helpers/http-helpers'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 export class GetUserByEmailController implements WebController {
   constructor (private readonly getUserByEmailUseCase: UseCase) {}
