+++ src/presentation/controllers/refresh-token/refresh-token.controller.ts
@@ -2,8 +2,8 @@ import type { WebController } from '@/core/presentation/web-controller'
 import type { HttpRequest, HttpResponse } from '@/infra/http/ports/http-protocol'
 import { ExpiredRefreshTokenError } from '@/application/usecases/refresh-token/errors/expired-refresh-token.error'
 import type { RefreshAccessTokenUseCase } from '@/application/usecases/refresh-token/refresh-token.usecase'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { badRequest, notFound, ok } from '@/presentation/helpers/http-helpers'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 export class RefreshAccessTokenController implements WebController {
   constructor (private readonly refreshTokenUseCase: RefreshAccessTokenUseCase) {}
